Extract module for soundcloud auth
