+++ src/sode/python/sode/soundcloud/cli.py
@@ -1,7 +1,7 @@
 from argparse import _SubParsersAction
 
 from sode.shared.cli.namespace import add_subcommand_parsers
-from sode.soundcloud.auth import add_auth
+from sode.soundcloud.authcmd import add_auth
 from sode.soundcloud.shared import SC_COMMAND
 from sode.soundcloud.thing import add_the_thing
 from sode.soundcloud.track import add_track
