+++ src/sode/python/sode/soundcloud/thing.py
@@ -3,11 +3,6 @@ import logging
 import os
 import textwrap
 from argparse import RawTextHelpFormatter, _SubParsersAction
-from typing import Any, Mapping
-
-from oauthlib.oauth2 import BackendApplicationClient
-from requests.auth import HTTPBasicAuth
-from requests_oauthlib import OAuth2Session
 
 from sode.shared.cli import argfactory, cmdfactory
 from sode.shared.cli.namespace import ProgramNamespace
@@ -16,6 +11,7 @@ from sode.shared.fp.either import Either, Left, Right
 from sode.shared.fp.option import Empty, Option, Value
 from sode.shared.oauth.token import AccessToken
 from sode.soundcloud import playlist
+from sode.soundcloud.authapi import scauth_fetch_tokens
 from sode.soundcloud.shared import SC_COMMAND
 
 logger = logging.getLogger(__name__)
@@ -136,55 +132,13 @@ def sodeenv_existing_access_token(args: ProgramNamespace) -> Option[AccessToken]
 ## SoundCloud auth
 
 
-class TokenResponse:
-    _mapping: dict[str, Any]
-    # access_token: str
-    # expires_at: float  # 1743781923.9585016
-    # expires_in: int  # 3599
-    # refresh_token: str
-    # scope: list[str]  # ['']
-    # token_type: str  # Bearer
-
-    def __init__(self, raw_response: Mapping[str, Any]):
-        self._mapping = dict(raw_response)
-
-    @staticmethod
-    def of(raw_response: Mapping[str, Any]) -> "TokenResponse":
-        return TokenResponse(raw_response)
-
-    @property
-    def access_token(self) -> Either[str, AccessToken]:
-        return AccessToken.expected(self._mapping["access_token"], self._mapping["token_type"])
-
-
 def scauth_authorize(args: ProgramNamespace) -> Either[str, AccessToken]:
     match sodeenv_existing_access_token(args):
         case Value(access_token):
             return Right(access_token)
         case Empty():
-            return scauth_fetch_tokens(args).flat_map(lambda response: response.access_token)
-
-
-def scauth_fetch_tokens(args: ProgramNamespace) -> Either[str, TokenResponse]:
-    """Authorize with the client_credentials workflow"""
-
-    logger.debug(
-        {
-            "fetch_tokens": {
-                "client_id": repr(args.client_id),
-                "token_url": repr(args.token_endpoint),
-            }
-        }
-    )
-
-    # https://developers.soundcloud.com/docs#authentication
-    auth = HTTPBasicAuth(args.client_id, args.client_secret)
-    client = BackendApplicationClient(client_id=args.client_id)
-    oauth = OAuth2Session(client=client)
-
-    try:
-        return Right(oauth.fetch_token(auth=auth, token_url=args.token_endpoint)).map(
-            lambda x: TokenResponse.of(x)
-        )
-    except Exception as err:
-        return Left(f"{args.token_endpoint}: {err}")
+            return scauth_fetch_tokens(
+                args.token_endpoint,
+                client_id=args.client_id,
+                client_secret=args.client_secret,
+            ).flat_map(lambda response: response.access_token)
